extract with pdf services and postprocess

+++ src/financial_tracker/__main__.py
@@ -1,8 +1,11 @@
 import json
 import logging
+import os
 import pathlib
 
 from financial_tracker.pdf import concat_pdfs
+from financial_tracker.pdf_services import PDFTextTableExtractor
+from financial_tracker.preprocess import PDFTextBuilder
 
 
 def main() -> None:
@@ -15,26 +18,40 @@ def main() -> None:
     globs = [list(dir.glob("*.pdf", case_sensitive=False)) for dir in dirs]
     cat_indices = sum([[i] * len(glob) for i, glob in enumerate(globs)], [])
     paths = [str(path) for glob in globs for path in glob]
-    statements_path = data_dir / "statements.pdf"
-    page_starts, page_ends = concat_pdfs(paths, str(statements_path))
-    logger.info(f"Concatenated statements saved to {statements_path}")
-    metadata_path = data_dir / "metadata.json"
-    with open(metadata_path, "w") as f:
-        json.dump(
-            {
-                path: {
-                    "page_start": page_start,
-                    "page_end": page_end,
-                    "category": categories[cat_index],
-                }
-                for path, page_start, page_end, cat_index in zip(
-                    paths, page_starts, page_ends, cat_indices
-                )
-            },
-            f,
-            indent=4,
+    statements_path = str(data_dir / "statements.pdf")
+    page_starts, page_ends = concat_pdfs(paths, statements_path)
+    logger.info(f"Concatenated statements saved to {statements_path=}")
+    metadata_path = str(data_dir / "metadata.json")
+    metadata = {
+        path: {
+            "page_start": page_start,
+            "page_end": page_end,
+            "category": categories[cat_index],
+        }
+        for path, page_start, page_end, cat_index in zip(
+            paths, page_starts, page_ends, cat_indices
         )
-    logger.info(f"Metadata saved to {metadata_path}")
+    }
+    extract_dir = str(data_dir / "extract")
+    if os.path.exists(metadata_path) and metadata == json.load(open(metadata_path)):
+        logger.info(f"Metadata already exists in {metadata_path=}")
+        logger.info(f"Loading extracted data from {extract_dir=}")
+    else:
+        with open(metadata_path, "w") as f:
+            json.dump(metadata, f, indent=4)
+        logger.info(f"Metadata saved to {metadata_path=}")
+        logger.info("Extracting data with Adobe PDF Services...")
+        extractor = PDFTextTableExtractor(
+            str(data_dir / "pdfservices-api-credentials.json")
+        )
+        extractor.extract_text_table(statements_path, extract_dir)
+        logger.info(f"Extracted data saved to {extract_dir=}")
+    pdf_text_builder = PDFTextBuilder(str(data_dir))
+    pdf_text = pdf_text_builder.process_pdf_data()
+    pdf_text_path = str(data_dir / "pdf_text.txt")
+    with open(pdf_text_path, "w") as f:
+        f.write(pdf_text)
+    logger.info(f"PDF text saved to {pdf_text_path=}")
 
 
 if __name__ == "__main__":
